test.py
import torch 
import cv2
import numpy as np

model = torch.hub.load('./yolov5/', 'custom', path='my_face_detection_model.pt', source='local') 

cap = cv2.VideoCapture(0)
while cap.isOpened():
    ret, frame = cap.read()
    
    # Make detections 
    results = model(frame)
    
    cv2.imshow('YOLO', np.squeeze(results.render()))
    
    if cv2.waitKey(10) & 0xFF == ord('q'):
        break
cap.release()
cv2.destroyAllWindows()